address null

+++ libraries/libft/ft_itoa_base.c
@@ -15,11 +15,11 @@ static int	ft_countdigits(unsigned long int integer, unsigned int base)
 
 char	*ft_itoa_base(unsigned long int number, unsigned int base)
 {
-	char			*converted_number;
-	size_t			thismanychars;
-	char			*ptr;
-	unsigned int	leftovers;
-	char			*representation;
+	char				*converted_number;
+	size_t				thismanychars;
+	char				*ptr;
+	unsigned long int	leftovers;
+	char				*representation;
 
 	representation = "0123456789abcdef";
 	thismanychars = ft_countdigits(number, base);
